fecth data 부분 수정
