+++ src/components/ui/skeleton.tsx
@@ -6,7 +6,7 @@ function Skeleton({
 }: React.HTMLAttributes<HTMLDivElement>) {
   return (
     <div
-      className={cn('animate-pulse rounded-md bg-gray-300', className)}
+      className={cn('animate-pulse rounded-md bg-gray-200', className)}
       {...props}
     />
   )
